libc: Drop stat/statfs 32->64 conversion code
New architectures, which do not define the ARCH_HAS_DEPRECATED_SYSCALLS
symbol, use the new interfaces from
libc/sysdeps/linux/common-no-legacy/bits/. Right now, only a new
struct statfs and strut stat are provided which have a similar layout
to the 64-bit equivalents but they use 32-bit members and a 32-bit
padding so they can be used in the 64-bit syscalls directly. This
eliminates the need to pass a 64-bit struct to the syscall which will
then be converted back to a 32-bit one. For dynamic loader, which only
uses the kernel structures, we define kernel_stat to be kernel_stat64
for these architectures so we can use the 64-bit syscalls directly.

See: http://lists.uclibc.org/pipermail/uclibc/2012-November/047255.html
(read thread for full context)

Reported-by: Rich Felker <dalias@aerifal.cx>
Signed-off-by: Markos Chandras <markos.chandras@imgtec.com>